# docs/swagger.yaml
definitions:
...
  config.File:
    properties:
      date:
        type: string
      height:
        type: string
      name:
        type: string
      size:
        type: string
      url:
        type: string
    type: object
...
  server.APIResponse:
    properties:
      code:
        type: integer
      data:
        description: omitempty를 사용하여 데이터가 없는 경우 필드를 생략
      message:
        type: string
    type: object
  server.Commission:
    properties:
      max_change_rate:
        type: string
      max_rate:
        type: string
      rate:
        type: string
    type: object
  server.Description:
    properties:
      details:
        type: string
      identity:
        type: string
      moniker:
        type: string
      website:
        type: string
    type: object
...
  server.Snapshots:
    properties:
      app:
        type: string
      base_url:
        type: string
      chain_id:
        type: string
      files:
        items:
          $ref: '#/definitions/config.File'
        type: array
      go_version:
        type: string
      img_url:
        type: string
      name:
        type: string
    type: object
...
  server.Uptime:
    properties:
      missed_block:
        type: integer
      percent:
        type: number
      tombstoned:
        type: boolean
      total_block:
        type: integer
    type: object
  server.Validator:
    properties:
      commission:
        $ref: '#/definitions/server.Commission'
      delegator_shares:
        type: string
      description:
        $ref: '#/definitions/server.Description'
      jailed:
        type: boolean
      operator_addr:
        type: string
      status:
        type: string
      tokens:
        type: string
      total_delegation_counts:
        type: string
      uptime:
        $ref: '#/definitions/server.Uptime'
      valcon_address:
        type: string
      voting_power: {}
      wallet_address:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /snapshot/{chain}:
    get:
      consumes:
      - application/json
      description: Get snapshot
      parameters:
      - description: Chain ID
        in: path
        name: chain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/server.Snapshots'
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/server.APIResponse'
            - properties:
                data:
                  type: object
              type: object
      summary: Get snapshot
      tags:
      - snapshots
...
  /validator/{chain}:
    get:
      consumes:
      - application/json
      description: Get validator
      parameters:
      - description: Chain ID
        in: path
        name: chain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/server.Validator'
              type: object
      summary: Get validator
      tags:
      - validators
...
swagger: "2.0"